chore: add raised handling error
